requirements.txt
django==3.1.5
djangorestframework==3.12.2
django-multiselectfield==0.1.12
django-registration-redux==2.9
Pillow==8.1.0
python-dotenv==0.15.0